pyproject.toml
[tool.poetry]
name = "nonebot-plugin-antiinsult"
version = "0.1.0"
description = "Anti-insult in NoneBot2"
authors = ["月ヶ瀬"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "nonebot_plugin_antiinsult" },
]
homepage = "https://github.com/tkgs0/nonebot-plugin-antiinsult"
repository = "https://github.com/tkgs0/nonebot-plugin-antiinsult"
keywords = ["nonebot", "anti-insult", "anti-abuse"]

[tool.poetry.dependencies]
python = "^3.8"
nonebot-adapter-onebot = "^2.1.5"
nonebot2 = "^2.0.0-rc.1"

[tool.poetry.dev-dependencies]

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true
implicit_reexport = true
pretty = true
show_error_codes = true
strict = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

